refactor: remove deprecated patch package and update database path configuration
