--- .changeset/config.json
@@ -1,11 +1,0 @@
-{
-  "$schema": "https://unpkg.com/@changesets/config@3.0.3/schema.json",
-  "changelog": "@changesets/cli/changelog",
-  "commit": false,
-  "fixed": [],
-  "linked": [],
-  "access": "restricted",
-  "baseBranch": "dev",
-  "updateInternalDependencies": "patch",
-  "ignore": []
-}
