--- .changeset/poor-nails-smoke.md
@@ -1,19 +1,0 @@
----
-"rusty-motors-shared-packets": patch
-"rusty-motors-transactions": patch
-"rusty-motors-database": patch
-"rusty-motors-gateway": patch
-"rusty-motors-personas": patch
-"rusty-motors-shared": patch
-"rusty-motors-lobby": patch
-"rusty-motors-login": patch
-"rusty-motors-mcots": patch
-"rusty-motors-patch": patch
-"rusty-motors-shard": patch
-"rusty-motors-cli": patch
-"rusty-motors-nps": patch
-"rusty-motors-socket": patch
-"rusty-motors-chat": patch
----
-
-Initial changeset
