--- .changeset/pre.json
@@ -1,24 +1,0 @@
-{
-  "mode": "pre",
-  "tag": "next",
-  "initialVersions": {
-    "rusty-motors-cli": "1.0.0-alpha1",
-    "rusty-motors-database": "1.0.0-alpha1",
-    "rusty-motors-gateway": "1.0.0-alpha1",
-    "rusty-motors-lobby": "1.0.0-alpha1",
-    "rusty-motors-login": "1.0.0-alpha1",
-    "rusty-motors-mcots": "1.0.0-alpha1",
-    "rusty-motors-nps": "1.0.0-alpha1",
-    "rusty-motors-patch": "1.0.0-alpha1",
-    "rusty-motors-personas": "1.0.0-alpha1",
-    "rusty-motors-shard": "1.0.0-alpha1",
-    "rusty-motors-shared": "1.0.0-alpha1",
-    "rusty-motors-shared-packets": "1.0.0-alpha1",
-    "rusty-motors-transactions": "1.0.0-alpha1",
-    "rusty-motors-chat": "1.0.0-alpha1",
-    "rusty-motors-socket": "1.0.0-alpha1"
-  },
-  "changesets": [
-    "poor-nails-smoke"
-  ]
-}
