+++ package.json
@@ -35,9 +35,6 @@
     "@databases/pg": "^5.5.0",
     "@dotenvx/dotenvx": "^1.32.0",
     "@fastify/sensible": "^6.0.2",
-    "@mikro-orm/core": "^6.4.2",
-    "@mikro-orm/postgresql": "^6.4.2",
-    "@mikro-orm/reflection": "^6.4.2",
     "@sentry/node": "^8.47.0",
     "@sentry/profiling-node": "8.47.0",
     "adminjs": "^7.8.13",
@@ -45,7 +42,6 @@
     "fastify": "^5.2.0",
     "moment": "^2.30.1",
     "pino-debug": "^2.0.0",
-    "pino-pretty": "^13.0.0",
     "rusty-motors-chat": "link:src/chat",
     "rusty-motors-cli": "link:packages/cli",
     "rusty-motors-database": "link:packages/database",
@@ -54,7 +50,6 @@
     "rusty-motors-login": "link:packages/login",
     "rusty-motors-mcots": "link:packages/mcots",
     "rusty-motors-nps": "link:packages/nps",
-    "rusty-motors-patch": "link:packages/patch",
     "rusty-motors-personas": "link:packages/persona",
     "rusty-motors-shard": "link:packages/shard",
     "rusty-motors-shared": "link:packages/shared",
@@ -65,7 +60,6 @@
   },
   "devDependencies": {
     "@biomejs/biome": "1.9.4",
-    "@changesets/cli": "^2.27.11",
     "@commitlint/cli": "^19.6.1",
     "@commitlint/config-conventional": "^19.6.0",
     "@databases/pg-migrations": "^5.0.3",
