--- packages/patch/CHANGELOG.md
@@ -1,7 +1,0 @@
-# rusty-motors-patch
-
-## 1.0.0-next.0
-
-### Patch Changes
-
-- ae13a4c: Initial changeset
