--- packages/patch/index.ts
@@ -1,1 +1,0 @@
-export { CastanetResponse } from "./src/PatchServer.js";
