--- packages/patch/package.json
@@ -1,31 +1,0 @@
-{
-	"name": "rusty-motors-patch",
-	"version": "1.0.0-next.0",
-	"exports": {
-		".": {
-			"import": "./index.js",
-			"require": "./index.js"
-		}
-	},
-	"type": "module",
-	"scripts": {
-		"check": "tsc",
-		"lint": "npx @biomejs/biome lint --write .",
-		"format": "npx @biomejs/biome format --write .",
-		"test": "vitest run --coverage"
-	},
-	"keywords": [],
-	"author": "",
-	"license": "AGPL-3.0",
-	"directories": {
-		"test": "test"
-	},
-	"description": "",
-	"devDependencies": {
-		"@vitest/coverage-v8": "2.1.8",
-		"vitest": "^2.1.8"
-	},
-	"dependencies": {
-		"@sentry/profiling-node": "8.47.0"
-	}
-}
