--- packages/patch/tsconfig.json
@@ -1,8 +1,0 @@
-{
-	"extends": "./../../tsconfig.base.json",
-	"compilerOptions": {
-		"incremental": true,
-		"composite": true
-	},
-	"include": ["index.ts", "src"]
-}
