+++ packages/transactions/test/getLobbies.test.ts
@@ -1,5 +1,4 @@
 import { describe, expect, it } from "vitest";
-import { mockPino } from "../../../test/factoryMocks.js";
 import { OldServerMessage } from "rusty-motors-shared";
 import { getLobbies } from "../src/getLobbies.js";
 
@@ -8,7 +7,6 @@ describe("getLobbies", () => {
 		// arrange
 		const connectionId = "1";
 		const packet = new OldServerMessage();
-		mockPino();
 
 		// act
 		const result = await getLobbies({
